refactor(socket-mode): simplify async connect lock acquire and expand tests
